fix: fixed rename issues
Fixed imports or renamed request/order model
Removed notification card
